Fixed issue #17469 [security]: SQL injection in Participants (editShareInfo)

+++ application/controllers/admin/participantsaction.php
@@ -1911,9 +1911,9 @@ $url .= "_view"; });
     public function editShareInfo()
     {
         $operation = Yii::app()->request->getPost('oper');
-        $shareIds = Yii::app()->request->getPost('id');
+        $shareIds = (int) Yii::app()->request->getPost('id');
         if ($operation == 'del') {
-// If operation is delete , it will delete, otherwise edit it
+            // If operation is delete , it will delete, otherwise edit it
             ParticipantShare::model()->deleteRow($shareIds);
         } else {
             $aData = array(
